requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
websockets>=16.0
openai>=1.0.0
openai-agents>=0.11.0
litellm>=1.0.0
httpx>=0.27.0
Pillow>=10.0.0
python-dotenv>=1.0.0
